requirements.txt
Flask==1.1.2
gunicorn==20.1.0
py-postgresql==1.2.2